% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcodes.R
\name{extract_barcode}
\alias{extract_barcode}
\title{INPUT = full file path
OUTPUT = Barcode extracted from ID1 field, if there is one there}
\usage{
extract_barcode(file)
}
\description{
INPUT = full file path
OUTPUT = Barcode extracted from ID1 field, if there is one there
}
