% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barcode_BMG.R
\name{get_barcode_BMG}
\alias{get_barcode_BMG}
\title{Extract plate barcodes from BMG plate reader data files}
\usage{
get_barcode_BMG(file, trim = FALSE)
}
\arguments{
\item{file}{File path to the plate whose barcode you wish to extract}

\item{trim}{If TRUE, trim leading 0s off of the extracted barcode (default = FALSE).}
}
\description{
Looks for a barcode present in the "ID1" field of plate reader data files
output by BMG Labtech instruments, such as FLUOstar, SPECTROstar, etc.
}
