% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tecan.R
\name{load_tecan}
\alias{load_tecan}
\title{Load and clean data from BMG plate readers}
\usage{
load_tecan(file, get.barcode = FALSE)
}
\arguments{
\item{file}{Raw-text file output from Tecan plate reader containing growth data}

\item{get.barcode}{If TRUE, will search for a plate barcode in the "ID1"
field of the file header.}
}
\description{
Loads in data from Tecan plate readers in a tidy (long) format.
Users can then directly plot the data or estimate growth parameters
with `grofitr()`.
}
