% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plates.R
\name{plot_plates}
\alias{plot_plates}
\title{Plot tidy plate reader data with ggplot2}
\usage{
plot_plates(data)
}
\arguments{
\item{data}{Plate reader data in tidy (long) format.}
}
\description{
Uses ggplot2 to plot an entire plate of data in a 12 x 8 grid.
}
