% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_dbl.R
\name{time_to_dbl}
\alias{time_to_dbl}
\title{Convert time to decimal}
\usage{
time_to_dbl(t)
}
\arguments{
\item{t}{A character vector of times in the form "HH h\[ MM min\]"}
}
\value{
A numeric vector of times in hour-decimal form
}
\description{
Converts time in the form "HH h MM min" to hours in decimal form
}
\examples{
time_to_dbl("12 h 1 min")
t <- c("0 h", "0 h 45 min", "1 h 30 min")
time_to_dbl(t)

}
\author{
Drew T. Doering, \email{dtdoering@wisc.edu}
}
